+++ apps/core-app/src/renderer/src/views/base/begin/internal/License.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts" name="License">
 import type { Component } from 'vue'
+import { useI18n } from 'vue-i18n'
 import Protocol from '~/assets/docs/license.md?raw'
 import AgreementTemplate from '~/components/addon/AgreementTemplate.vue'
 import AccountDo from './AccountDo.vue'
@@ -8,6 +9,7 @@ import Forbidden from './Forbidden.vue'
 type StepFunction = (call: { comp: Component; rect?: { width: number; height: number } }) => void
 
 const step: StepFunction = inject('step')!
+const { t } = useI18n()
 
 function handleAgree(val: boolean): void {
   step({
@@ -18,7 +20,11 @@ function handleAgree(val: boolean): void {
 
 <template>
   <div class="License">
-    <AgreementTemplate title="License" :agree="handleAgree" :agreement="Protocol" />
+    <AgreementTemplate
+      :title="t('beginner.license.title')"
+      :agree="handleAgree"
+      :agreement="Protocol"
+    />
   </div>
 </template>